+++ client/src/scripts/globals.ts
@@ -1,5 +1,6 @@
 export const DEPTH = {
   body: 100,
+  particles: 101,
   ground: -1,
   backgroundEffect: 90,
   tile: 20,