+++ client/src/scripts/network/bodySynchronizer.ts
@@ -1,15 +1,38 @@
 import { BodySchema } from "../generated/BodySchema";
+import { Texture } from "../generated/Texture";
 import { DEPTH } from "../globals";
 import { createLogger } from "../logger";
-import BodySprite from "../objects/Body";
 import LevelScene from "../scenes/LevelScene";
+import { getTextureFrameKey } from "../shared/schema-utils";
 const log = createLogger("client:body-synchronizer");
 
 export function bodySynchronizer(scene: LevelScene, body: BodySchema) {
   log("added new body with id", body.id);
 
+  if (scene.network.ownPlayer?.bodyId === body.id) {
+    log("Adding body of our player!");
+  }
+
   // adding new player to scene
-  const bodySprite = new BodySprite(scene, 200, 200);
+  const bodySprite = scene.add.image(
+    200,
+    200,
+    body.texture.key,
+    getTextureFrameKey(body.texture)
+  );
+  bodySprite.setDepth(DEPTH.body);
+
+  scene.levelListeners.push(
+    body.listen("texture", (texture) => {
+      log(
+        "Setting body texture=%s, frame=%s",
+        texture.key,
+        getTextureFrameKey(texture)
+      );
+      bodySprite.setTexture(texture.key, getTextureFrameKey(texture));
+    })
+  );
+  // bodySprite.setTexture(body.texture.key, getTextureFrameKey(body.texture));
 
   scene.levelListeners.push(
     body.position.listen("x", (x) => bodySprite.setX(x))
@@ -43,6 +66,23 @@ export function bodySynchronizer(scene: LevelScene, body: BodySchema) {
     log("player", body.id, "removed");
     bodySprite.destroy();
     destroyParticles();
+
+    log(
+      "Setting body texture=%s, frame=%s",
+      body.texture.key,
+      getTextureFrameKey(body.texture)
+    );
+
+    scene.particles.emitDeathExplosion(
+      body.position.x,
+      body.position.y,
+      body.texture.key,
+      getTextureFrameKey(body.texture)
+    );
+
+    if (scene.network.ownPlayer?.bodyId === body.id) {
+      log("Removing body of our player!");
+    }
   };
 
   scene.levelListeners.push(