+++ server/src/game/BodyFactory.ts
@@ -1,4 +1,4 @@
-import { BodySchema } from "./schema/GameState";
+import { BodySchema, Texture } from "./schema/GameState";
 import { gameEnvironment } from "./gameEnvironment";
 
 export class BodyFactory {
@@ -10,6 +10,7 @@ export class BodyFactory {
     body.position.x = 50;
     body.position.y = 50;
     body.energy = 0;
+    body.texture = new Texture("body.blue.1");
     body.maxEnergy = gameEnvironment.normalEnergy;
     body.speed = gameEnvironment.speedNormal;
     body.isSprinting = false;